Fix GDN/KDA bugs, require CUDA kernels, add cache-aware tests (#451)
