+++ fast_llm_external_models/tests/test_apriel2/test_cache_apriel2_specific.py
@@ -18,7 +18,7 @@ Fixtures used from conftest.py:
 import pytest
 import torch
 
-from fast_llm_external_models.apriel2.cache import Apriel2Cache
+from fast_llm_external_models.apriel2.modeling_apriel2 import Apriel2Cache
 
 # =============================================================================
 # STOCHASTIC MIXER ROUTING
