+++ fast_llm_external_models/tests/test_apriel2/test_cache_contracts.py
@@ -27,7 +27,7 @@ Fixtures used from conftest.py:
 import pytest
 import torch
 
-from fast_llm_external_models.apriel2.cache import Apriel2Cache, _AttentionCache
+from fast_llm_external_models.apriel2.modeling_apriel2 import Apriel2Cache, _AttentionCache
 
 # =============================================================================
 # SECTION 1: FULL ATTENTION - _AttentionCache vs DynamicLayer
